Elaborate on subsetting in the README
This follows up on #58 which I opened because I wondered what the term "subsetting" in the README was about. I learned a lot from the discussion in #58, and I don't think the README needs to explain everything, yet the term appears often (seven times) and prominent enough to justify a superficial explanation.